fix: bound build execution and clean temporary output

+++ src/Adapters/BuildService/Atropos.Adapters.BuildService.pas
@@ -7,12 +7,14 @@ uses
 type
   IBuildProcessRunner = interface
     ['{A44BDD56-4A86-46C5-A2B4-C95A6B2D54E2}']
-    function Execute(const ACommand: string; out AOutput: string; out AExitCode: Cardinal): Boolean;
+    function Execute(const ACommand: string; ATimeoutMs: Cardinal; out AOutput: string;
+      out AExitCode: Cardinal; out ATimedOut: Boolean): Boolean;
   end;
 
   TWin32BuildProcessRunner = class(TInterfacedObject, IBuildProcessRunner)
   public
-    function Execute(const ACommand: string; out AOutput: string; out AExitCode: Cardinal): Boolean;
+    function Execute(const ACommand: string; ATimeoutMs: Cardinal; out AOutput: string;
+      out AExitCode: Cardinal; out ATimedOut: Boolean): Boolean;
   end;
 
   TBuildOutputParser = class
@@ -25,29 +27,31 @@ type
     FEnvService: IDelphiEnvironmentService;
     FLogger: ILogger;
     FProcessRunner: IBuildProcessRunner;
+    FTimeoutMs: Cardinal;
     function GetDelphiFriendlyName(const ADelphiPath: string): string;
   public
     constructor Create(AEnvService: IDelphiEnvironmentService; ALogger: ILogger = nil;
-      AProcessRunner: IBuildProcessRunner = nil);
+      AProcessRunner: IBuildProcessRunner = nil; ATimeoutMs: Cardinal = 600000);
     function BuildProject(const AProjectPath: string): TBuildMetrics;
   end;
 
 implementation
-uses System.Classes, System.Generics.Collections, System.IOUtils, System.RegularExpressions, Winapi.Windows,
+uses System.Classes, System.Generics.Collections, System.IOUtils, System.Math, System.RegularExpressions, Winapi.Windows,
   System.SysUtils;
 
 constructor TBuildServiceAdapter.Create(AEnvService: IDelphiEnvironmentService; ALogger: ILogger;
-  AProcessRunner: IBuildProcessRunner);
+  AProcessRunner: IBuildProcessRunner; ATimeoutMs: Cardinal);
 begin
   FEnvService := AEnvService;
   FLogger := ALogger;
   FProcessRunner := AProcessRunner;
   if not Assigned(FProcessRunner) then
     FProcessRunner := TWin32BuildProcessRunner.Create;
+  FTimeoutMs := ATimeoutMs;
 end;
 
-function TWin32BuildProcessRunner.Execute(const ACommand: string; out AOutput: string;
-  out AExitCode: Cardinal): Boolean;
+function TWin32BuildProcessRunner.Execute(const ACommand: string; ATimeoutMs: Cardinal;
+  out AOutput: string; out AExitCode: Cardinal; out ATimedOut: Boolean): Boolean;
 var
   LSecurityAttributes: TSecurityAttributes;
   LReadPipe: THandle;
@@ -56,12 +60,16 @@ var
   LProcessInfo: TProcessInformation;
   LBuffer: array[0..4095] of AnsiChar;
   LBytesRead: DWORD;
+  LBytesAvailable: DWORD;
   LOutputStream: TStringStream;
   LMutableCmd: string;
+  LStartTick: UInt64;
+  LWaitResult: DWORD;
 begin
   Result := False;
   AOutput := EmptyStr;
   AExitCode := Cardinal(-1);
+  ATimedOut := False;
 
   LSecurityAttributes.nLength := SizeOf(TSecurityAttributes);
   LSecurityAttributes.bInheritHandle := True;
@@ -88,19 +96,37 @@ begin
 
       LOutputStream := TStringStream.Create('');
       try
-        while ReadFile(LReadPipe, LBuffer, SizeOf(LBuffer) - 1, LBytesRead, nil) and (LBytesRead > 0) do
+        LStartTick := GetTickCount64;
+        repeat
         begin
-          LBuffer[LBytesRead] := #0;
+          while PeekNamedPipe(LReadPipe, nil, 0, nil, @LBytesAvailable, nil) and
+            (LBytesAvailable > 0) do
+          begin
+            if not ReadFile(LReadPipe, LBuffer, Min(Cardinal(SizeOf(LBuffer) - 1), LBytesAvailable), LBytesRead, nil) then
+              Break;
+            LOutputStream.WriteBuffer(LBuffer, LBytesRead);
+          end;
+          LWaitResult := WaitForSingleObject(LProcessInfo.hProcess, 10);
+          if LWaitResult = WAIT_OBJECT_0 then
+            Break;
+          if (ATimeoutMs > 0) and (GetTickCount64 - LStartTick >= ATimeoutMs) then
+          begin
+            ATimedOut := True;
+            TerminateProcess(LProcessInfo.hProcess, ERROR_TIMEOUT);
+            WaitForSingleObject(LProcessInfo.hProcess, 5000);
+            Break;
+          end;
+        end
+        until False;
+        while ReadFile(LReadPipe, LBuffer, SizeOf(LBuffer) - 1, LBytesRead, nil) and (LBytesRead > 0) do
           LOutputStream.WriteBuffer(LBuffer, LBytesRead);
-        end;
         AOutput := LOutputStream.DataString;
       finally
         LOutputStream.Free;
       end;
 
-      WaitForSingleObject(LProcessInfo.hProcess, INFINITE);
       if GetExitCodeProcess(LProcessInfo.hProcess, AExitCode) then
-        Result := True;
+        Result := not ATimedOut;
       CloseHandle(LProcessInfo.hProcess);
       CloseHandle(LProcessInfo.hThread);
     end;
@@ -205,6 +231,7 @@ var
   LOutput: string;
   LStartTick: UInt64;
   LExitCode: Cardinal;
+  LTimedOut: Boolean;
 begin
   Result := Default(TBuildMetrics);
   if not Assigned(FEnvService) then
@@ -233,25 +260,32 @@ begin
   if Assigned(FLogger) then FLogger.Log('Executing Build via bds.exe (Universal Compiler): ' + LBdsCmd);
 
   LStartTick := GetTickCount64;
-  if not FProcessRunner.Execute(LBdsCmd, LOutput, LExitCode) then
-  begin
-    Result.Success := False;
-    Result.ErrorMessage := 'Failed to execute bds.exe process.';
-    Exit;
+  try
+    if not FProcessRunner.Execute(LBdsCmd, FTimeoutMs, LOutput, LExitCode, LTimedOut) then
+    begin
+      Result.Success := False;
+      if LTimedOut then
+        Result.ErrorMessage := Format('bds.exe build timed out after %d ms.', [FTimeoutMs])
+      else
+        Result.ErrorMessage := 'Failed to execute bds.exe process.';
+      Exit;
+    end;
+
+    if not TFile.Exists(LErrFile) then
+    begin
+      Result.Success := False;
+      Result.ErrorMessage := 'Failed to read bds.exe error file output.';
+      Exit;
+    end;
+
+    LOutput := TFile.ReadAllText(LErrFile);
+    Result := TBuildOutputParser.Parse(LOutput, AProjectPath, LExitCode);
+    Result.DelphiVersion := GetDelphiFriendlyName(LDelphiPath);
+    Result.CompileTimeMs := Int64(GetTickCount64 - LStartTick);
+  finally
+    if TFile.Exists(LErrFile) then
+      TFile.Delete(LErrFile);
   end;
-  
-  if not TFile.Exists(LErrFile) then
-  begin
-    Result.Success := False;
-    Result.ErrorMessage := 'Failed to read bds.exe error file output.';
-    Exit;
-  end;
-  
-  LOutput := TFile.ReadAllText(LErrFile);
-  TFile.Delete(LErrFile);
-  Result := TBuildOutputParser.Parse(LOutput, AProjectPath, LExitCode);
-  Result.DelphiVersion := GetDelphiFriendlyName(LDelphiPath);
-  Result.CompileTimeMs := Int64(GetTickCount64 - LStartTick);
 end;
 
 end.
